API

JuliaBUGS.Parser.@bugs — Macro

@bugs(program::Expr)
@bugs(program::String; replace_period::Bool=true, no_enclosure::Bool=false)

Constructs a Julia Abstract Syntax Tree (AST) representation of a BUGS program. This macro supports two forms of input: a Julia expression or a string containing the BUGS program code.

When provided with a string, the macro parses it as a BUGS program, with optional arguments to control parsing behavior.

When given an expression, it performs syntactic checks to ensure compatibility with BUGS syntax.

Arguments for String Input

For the string input variant, the following optional arguments are available:

replace_period::Bool: When set to true, all periods (.) in the BUGS code are replaced. This is enabled by default.

no_enclosure::Bool: When true, the parser does not require the BUGS program to be enclosed within model{ ... } brackets. By default, this is set to false.

JuliaBUGS.compile — Function

compile(model_def[, data, initializations])

Compile a BUGS model into a log density problem.

Arguments

model_def::Expr: The BUGS model definition.

data::NamedTuple or AbstractDict: The data to be used in the model. If none is passed, the data will be assumed to be empty.

initializations::NamedTuple or AbstractDict: The initial values for the model parameters. If none is passed, the parameters will be assumed to be initialized to zero.

is_transformed::Bool=true: If true, the model parameters during inference will be transformed to the unconstrained space.

Returns

A BUGSModel object representing the compiled model.